Using MathF in float classes

+++ src/Math/Core/double2.cs
@@ -107,7 +107,7 @@ namespace Fusee.Math.Core
         /// The length.
         /// </value>
         /// <see cref="LengthSquared" />
-        public double Length => (double)System.Math.Sqrt(LengthSquared);
+        public double Length => System.Math.Sqrt(LengthSquared);
 
         #endregion public double Length
